move CosmosMetadata enum to root level as public api
